robonomics messages validation improvements
